==> scripts/13_yelp_seeds.py <==
from gensim.models import Word2Vec
import numpy as np
import pandas as pd
from src.utils.seed_words import all_seeds


w2vec_model = Word2Vec.load('data/raw/yelp_w2vec/yelp_w2vec.model')
vocabs = set(w2vec_model.wv.index2entity) # 363658 amazon: 43750

positive_seeds, negative_seeds = all_seeds() # 52, 58
positive_seeds = [w for w in positive_seeds if w in vocabs] # 31
negative_seeds = [w for w in negative_seeds if w in vocabs] # 35

positive_vectors = np.concatenate([w2vec_model[w].reshape(1,-1) for w in positive_seeds])
negative_vectors = np.concatenate([w2vec_model[w].reshape(1,-1) for w in negative_seeds])

for idx in range(len(positive_seeds)):
    assert (w2vec_model[positive_seeds[idx]] == positive_vectors[idx]).all()

for idx in range(len(negative_seeds)):
    assert (w2vec_model[negative_seeds[idx]] == negative_vectors[idx]).all()

train_df = pd.DataFrame(np.concatenate([positive_vectors, negative_vectors]),
                        index=pd.Index(positive_seeds+negative_seeds, name='SEED')
                        )
train_df['LABELS'] = [0 if w in negative_seeds else 1 for w in positive_seeds+negative_seeds]

assert (train_df[list(range(500))].values == np.concatenate([positive_vectors, negative_vectors])).all()

train_df = train_df.sample(frac=1, random_state=42)

train_df.to_csv('data/processed/yelp_seed_vectors_13.csv')
